+++ packages/plugin-packages/mashroom-portal/src/backend/utils/render_utils.ts
@@ -1,16 +1,14 @@
+
 import {PORTAL_PAGE_TEMPLATE_NAME, PORTAL_APP_WRAPPER_TEMPLATE_NAME, PORTAL_APP_ERROR_TEMPLATE_NAME} from '../constants';
 import minimalTemplatePortal from '../theme/minimal_template_portal';
 import defaultTemplateAppWrapper from '../theme/default_template_app_wrapper';
 import defaultTemplateAppError from '../theme/default_template_app_error';
+import {renderServerSide} from './ssr_utils';
 
 import type {Request, Response} from 'express';
 import type {MashroomLogger} from '@mashroom/mashroom/type-definitions';
-import type {MashroomPortalPageRenderModel} from '../../../type-definitions';
-import {
-    MashroomPortalAppErrorRenderModel,
-    MashroomPortalAppWrapperRenderModel,
-} from '../../../type-definitions';
-import type {MashroomPortalPageAppsInfo} from '../../../type-definitions/internal';
+import type {MashroomPortalPageRenderModel, MashroomPortalAppErrorRenderModel, MashroomPortalAppWrapperRenderModel} from '../../../type-definitions';
+import type {MashroomPortalPageAppsInfo, MashroomPortalPageContentRenderResult} from '../../../type-definitions/internal';
 
 export const renderPage = async (themeExists: boolean, model: MashroomPortalPageRenderModel, req: Request, res: Response, logger: MashroomLogger): Promise<string> => {
     const fallback = () => minimalTemplatePortal(model);
@@ -67,23 +65,28 @@ export const renderAppErrorToClientTemplate = async (themeExists: boolean, messa
     return renderAppError(themeExists, model, req, res, logger);
 }
 
-export const renderPageContent = async (portalLayout: string, portalAppInfo: MashroomPortalPageAppsInfo, themeExists: boolean, messages: (key: string) => string, req: Request, res: Response, logger: MashroomLogger): Promise<string> => {
-    let portalContent = portalLayout;
+export const renderPageContent = async (portalLayout: string, portalAppInfo: MashroomPortalPageAppsInfo, themeExists: boolean, messages: (key: string) => string, req: Request, res: Response, logger: MashroomLogger): Promise<MashroomPortalPageContentRenderResult> => {
+    const serverSideRenderedApps: Array<string> = [];
     const appAreas = Object.keys(portalAppInfo);
     const promises: Array<Promise<Array<string>>> = [];
+
+    let pageContent = portalLayout;
     for (const appAreaId of appAreas) {
         promises.push(
             Promise.all(
-                portalAppInfo[appAreaId].map(({pluginName, appSetup}) => {
+                portalAppInfo[appAreaId].map(async ({pluginName, appSetup}) => {
+                    const {appId, title} = appSetup;
 
-                    // TODO: check if app supports SSR
-                    const appSSRHtml = null;
+                    const appSSRHtml = await renderServerSide(pluginName, appSetup, req, logger);
+                    if (appSSRHtml && serverSideRenderedApps.indexOf(pluginName) === -1) {
+                        serverSideRenderedApps.push(pluginName);
+                    }
 
                     const model: MashroomPortalAppWrapperRenderModel = {
-                        appId: appSetup.appId,
+                        appId,
                         pluginName,
                         safePluginName: getSafePluginName(pluginName),
-                        title: appSetup.title || pluginName,
+                        title: title || pluginName,
                         messages,
                         appSSRHtml,
                     };
@@ -92,7 +95,7 @@ export const renderPageContent = async (portalLayout: string, portalAppInfo: Mas
             ).then(
                 (result) => result,
                 (error) => {
-                    logger.error(`Rendering apps in appArea ${appAreaId} failed`, error);
+                    logger.error(`Rendering apps in appArea '${appAreaId}' failed`, error);
                     return [];
                 }
             )
@@ -104,13 +107,16 @@ export const renderPageContent = async (portalLayout: string, portalAppInfo: Mas
        const appHtmls = appAreasAppHtmls[idx];
        if (appHtmls.length > 0) {
            const appHtml = appHtmls.join('');
-           const startIdxArea = portalContent.search(`<[a-zA-Z]+.*id=["']${appAreaId}["']`);
-           const beginAreaContent = portalContent.indexOf('>', startIdxArea) + 1;
-           portalContent = [portalContent.slice(0, beginAreaContent), appHtml, portalContent.slice(beginAreaContent)].join('');
+           const startIdxArea = pageContent.search(`<[a-zA-Z]+.*id=["']${appAreaId}["']`);
+           const beginAreaContent = pageContent.indexOf('>', startIdxArea) + 1;
+           pageContent = [pageContent.slice(0, beginAreaContent), appHtml, pageContent.slice(beginAreaContent)].join('');
        }
     });
 
-    return portalContent;
+    return {
+        pageContent,
+        serverSideRenderedApps,
+    };
 }
 
 const renderToString = (template: string, templateMustExist: boolean, model: any, fallback: () => string, res: Response, logger: MashroomLogger): Promise<string> => {