+++ packages/plugin-packages/mashroom-portal/src/frontend/welcome-app/js/index.ts
@@ -4,7 +4,9 @@ import template from './template';
 import type {MashroomPortalAppPluginBootstrapFunction} from '../../../../type-definitions';
 
 const bootstrap: MashroomPortalAppPluginBootstrapFunction = (element, portalAppSetup) => {
-    element.innerHTML = template(portalAppSetup.resourcesBasePath);
+    if (!element.querySelector('.welcome-app-content')) {
+        element.innerHTML = template(portalAppSetup.resourcesBasePath);
+    }
     return Promise.resolve();
 };