+++ packages/plugin-packages/mashroom-portal/src/frontend/welcome-app/js/indexSSR.ts
@@ -4,7 +4,7 @@ import template from './template';
 import type {MashroomPortalAppPluginSSRBootstrapFunction} from '../../../../type-definitions';
 
 const bootstrap: MashroomPortalAppPluginSSRBootstrapFunction = async (portalAppSetup) => {
-    return template(portalAppSetup.resourcesBasePath);
+    return template(portalAppSetup.resourcesBasePath, true);
 };
 
 export default bootstrap;