+++ packages/plugin-packages/mashroom-portal/src/frontend/welcome-app/js/template.ts
@@ -1,17 +1,17 @@
 
-import logoImg from '../assets/mashroom_logo.png';
-
-export default (resourcesBasePath: string) => `
+export default (resourcesBasePath: string, ssr = false) => `
     <div class="mashroom-welcome-portal-app">
         <h4>Welcome to the Mashroom Portal Server!</h4>
         <div class="welcome-app-content">
-            <img src="${resourcesBasePath}/${logoImg}"/>
+            <img src="${resourcesBasePath}/welcome-app/assets/mashroom_logo.png"/>
             <p>
-                This demonstration portal page integrates multiple standalone (single page) apps.
+                This demo portal page integrates multiple standalone (single page) applications.
                 <br/>
                 All of them are written in different frontend technologies.
                 <br/>
                 They make use of the frontend messaging bus to communicate with each other (Ping button!).
+                <br />
+                <small>(This has rendered on the ${ssr ? 'server-side' : 'client-side'})</small>
             </p>
         </div>
     </div>