+++ packages/plugin-packages/mashroom-portal/test/backend/controllers/PortalPageController.test.ts
@@ -12,7 +12,13 @@ setPortalPluginConfig({
     defaultLayout: 'foo',
     warnBeforeAuthenticationExpiresSec: 120,
     autoExtendAuthentication: false,
-    defaultProxyConfig: {}
+    defaultProxyConfig: {},
+    sslConfig: {
+        sslEnabled: false,
+        renderTimoutMs: 2000,
+        cacheTTLSec: 300,
+        inlineStyles: true,
+    }
 });
 
 const theme: MashroomPortalTheme = {