+++ packages/plugin-packages/mashroom-portal/test/backend/utils/ssr_bootstrap.js
@@ -1,0 +1,7 @@
+
+var bootstrap = function bootstrap(portalAppSetup) {
+    return Promise.resolve('this is a test');
+};
+
+var _default = bootstrap;
+exports.default = _default;